Update and round out .*
Add cases for vectors and matrices being mixed, though now I have to rethink
if they belong here, or the whole vectors directory should just be merged in
as a subset of matrices.

Also, rewrite the functions with the new (with-result) and (do-matrix)
macros, as well as (mref), which is how I'd imagined some of these would work
originally.